Implement client/master REST CRUD interface.
Clients (compile, link wrappers, reporters, ...) can use the interface
to interoperate with the master data model. All CRUD operations are
implemented for source, dependency and target entities. A test was
added under integration/ that expects a master to be available, and
then performs all operations in a test process. Run a master, and then
execute 'go test' in the integration/ directory. The test is also a
good reference on how to implement client/master communication.