QMSTR build within docker
Provide Dockerfile to build a docker image with a working qmstr
installation.
Containers can mount the qmstr source repository to /qmstr and run the
container with `dev build` to build qmstr in the container.
If you mount a project's source code to /build you can build your project
with qmstr in the container by running it with e.g. make.