Add config step running scancode
Since scancode's analysis is way too slow to be called from every
compile step a pre-scan is inplemented.
After the master server is started a configuration containing the
working directory is passed the /config endpoint. This will result in a
scancode scan of that very directory. The status can be checked via the
/health endpoint. When scanned: true is returned as part of the result
of the health endpoint the compilation can start and the wrapper can
start submitting source files.